fix: acknowledge queued RRD writes and supervise Boost children

+++ poller_realtime.php
@@ -136,7 +136,11 @@ if ($rrdtool_pipe === false) {
 }
 
 /* process poller output */
-process_poller_output_rt($rrdtool_pipe, $poller_id, $interval);
+if (process_poller_output_rt($rrdtool_pipe, $poller_id, $interval) === false) {
+	rrd_close($rrdtool_pipe);
+	db_close();
+	exit(1);
+}
 
 /* close rrd */
 rrd_close($rrdtool_pipe);
@@ -191,6 +195,7 @@ function process_poller_output_rt($rrdtool_pipe, $poller_id, $interval) {
 		ON dl.id = port.local_data_id
 		WHERE port.poller_id = ?',
 		array($poller_id));
+	if ($results === false) { return false; }
 
 	if (cacti_sizeof($results)) {
 		/* create an array keyed off of each .rrd file */
@@ -276,6 +281,11 @@ function process_poller_output_rt($rrdtool_pipe, $poller_id, $interval) {
 			}
 		}
 
+		$rrds_processed = rrdtool_function_update($rrd_update_array, $rrdtool_pipe);
+		if ($rrds_processed === false) {
+			return false;
+		}
+
 		/* make sure each .rrd file has complete data */
 		foreach ($results as $item) {
 			db_execute_prepared('DELETE FROM poller_output_realtime
@@ -286,7 +296,7 @@ function process_poller_output_rt($rrdtool_pipe, $poller_id, $interval) {
 				array($item['local_data_id'], $item['rrd_name'], $item['time'], $poller_id));
 		}
 
-		$rrds_processed = rrdtool_function_update($rrd_update_array, $rrdtool_pipe);
+
 	}
 
 	return $rrds_processed;
